late git hub fixes

+++ OpenGL 3D - William de Beer/Cloth.cpp
@@ -275,8 +275,8 @@ void Cloth::Update(float _dT, CCamera* _camera)
 				// Apply constraints to node
 				m_Nodes[i * m_Height + j]->ApplyConstraint(m_Nodes[i * m_Height + j]->GetConnection(Side::TOP), m_Spacing);
 				m_Nodes[i * m_Height + j]->ApplyConstraint(m_Nodes[i * m_Height + j]->GetConnection(Side::RIGHT), m_Spacing);
-				m_Nodes[i * m_Height + j]->ApplyConstraint(m_Nodes[i * m_Height + j]->GetConnection(Side::TR), sqrtf(pow(m_Spacing, 2) * 2.0f));
-				m_Nodes[i * m_Height + j]->ApplyConstraint(m_Nodes[i * m_Height + j]->GetConnection(Side::BR), sqrtf(pow(m_Spacing, 2) * 2.0f));
+				m_Nodes[i * m_Height + j]->ApplyConstraint(m_Nodes[i * m_Height + j]->GetConnection(Side::TR), sqrtf(powf(m_Spacing, 2) * 2.0f));
+				m_Nodes[i * m_Height + j]->ApplyConstraint(m_Nodes[i * m_Height + j]->GetConnection(Side::BR), sqrtf(powf(m_Spacing, 2) * 2.0f));
 			
 				// Check if it is to be destroyed and destroy it.
 				if (m_Nodes[i * m_Height + j]->GetToBeDestroyed())
